requirements.txt
# Python dependencies for Ethereum Mempool Analysis

# Data processing
pandas>=2.0.0
pyarrow>=12.0.0
numpy>=1.24.0

# ClickHouse client
clickhouse-driver>=0.2.6
requests>=2.28.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
notebook>=6.5.0

# DuckDB for fast local analysis (optional)
duckdb>=0.9.0
